Implement back-off in workers
Implement back-off in workers. Previously, when TaskAction had an 
exception (for example CQL query failed) a retry TaskAction would be
queued instantly, which could cause a big load on the Scylla cluster.

This change adds a back-off in such case. The chosen back-off strategy
is an exponential back-off with jitter (10ms doubling up to 30 seconds).