Add withStreamId, withTimeMs to MockRawChange
Add two utility methods to MockRawChange.Builder, that makes
it easier to construct MockRawChange. withStreamId finds a stream id
in the provided generation metadata, while withTimeMs generates
a TIMEUUID from a given epoch timestamp. The generated TIMEUUID
is in the middle of millisecond unit, so that there is no conflict
with startOf and endOf TIMEUUIDs.