Rework of Driver3SchemaBuilder
This commit is essentially a complete rework of Driver3SchemaBuilder.
This rework was made to fix potential problems with previous
Driver3SchemaBuilder, which could potentially return incorrect schema
when base table schema or CDC log schema returned by the driver was
stale.

The new Driver3SchemaFactory (it was renamed) uses the schema from both
tables, but relies as much as possible on CDC log schema. The returned
schema is always consistent with CDC log schema (or an exception is thrown).

Refs #45.
Refs scylladb/scylla#7824.