Wait for the first generation to start in ITs
Wait for the first generation to start in BaseScyllaIntegrationTest.
In Scylla 4.3, the first added generation is in the future (added
ring delays). Therefore, when INSERTing rows you could get a
"could not find any CDC stream" error. Waiting for the first generation
fixes this problem. In case of Scylla 4.4+, this workaround is not
needed as the delay is not added.